+++ projects/egeo-demo/src/app/demos/st-form-demo/form-editor/st-form-editor-demo.ts
@@ -56,4 +56,8 @@ export class StFormEditorDemoComponent {
          }, 1000
       );
    }
+
+   onClickLink(fieldPath: string): void {
+      console.log(fieldPath);
+   }
 }