+++ projects/egeo/src/lib/st-form/spec/resources/json-schema.ts
@@ -8,16 +8,17 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import { JSONSchema4 } from 'json-schema';
 
-export const JSON_SCHEMA: JSONSchema4 = {
+import { StFormSchema } from '../../st-form.model';
+
+export const JSON_SCHEMA: StFormSchema = {
    '$schema': 'http://json-schema.org/schema#',
    'title': 'General',
    'properties': {
       'genericIntegerInput': {
          'title': 'Generic number',
          'description': 'Generic input description',
-         'type': 'integer',
+         'type': <any> 'integer',
          'default': 5,
          'minimum': 6,
          'maximum': 10,
@@ -28,7 +29,7 @@ export const JSON_SCHEMA: JSONSchema4 = {
       'genericLongText': {
          'title': 'Generic long text',
          'description': 'Generic long text description',
-         'type': 'string',
+         'type': <any> 'string',
          'minLength': 2,
          'maxLength': 200,
          'examples': ['This is a long text and can be longer']
@@ -36,7 +37,7 @@ export const JSON_SCHEMA: JSONSchema4 = {
       'genericNumberInput': {
          'title': 'Generic number',
          'description': 'Generic input description',
-         'type': 'number',
+         'type': <any> 'number',
          'default': 1,
          'minimum': 0.5,
          'maximum': 10,
@@ -47,7 +48,7 @@ export const JSON_SCHEMA: JSONSchema4 = {
       'requiredNumber': {
          'title': 'Required number',
          'description': 'Required input description',
-         'type': 'integer',
+         'type': <any> 'integer',
          'default': 5,
          'examples': [6]
 
@@ -55,7 +56,7 @@ export const JSON_SCHEMA: JSONSchema4 = {
       'minNumber': {
          'title': 'Min number',
          'description': 'Min number input description',
-         'type': 'number',
+         'type': <any> 'number',
          'default': 8.6,
          'minimum': 6,
          'examples': [9.2]
@@ -64,14 +65,14 @@ export const JSON_SCHEMA: JSONSchema4 = {
       'maxNumber': {
          'title': 'Max number',
          'description': 'Max number input description',
-         'type': 'number',
+         'type': <any> 'number',
          'default': 7.6,
          'maximum': 6
       },
       'minAndMaxNumber': {
          'title': 'Number in a range',
          'description': 'This number has to be between 7 and 19',
-         'type': 'number',
+         'type': <any> 'number',
          'default': 8.6,
          'minimum': 6,
          'maximum': 10,
@@ -81,7 +82,7 @@ export const JSON_SCHEMA: JSONSchema4 = {
       'genericTextInput': {
          'title': 'Required text',
          'description': 'This is a generic text',
-         'type': 'string',
+         'type': <any> 'string',
          'default': 'Generic text',
          'minLength': 2,
          'maxLength': 6,
@@ -90,27 +91,27 @@ export const JSON_SCHEMA: JSONSchema4 = {
       'requiredText': {
          'title': 'Required text',
          'description': 'This is a required text',
-         'type': 'string',
+         'type': <any> 'string',
          'default': '5'
       },
       'minLengthText': {
          'title': 'Text with a max length',
          'description': 'You have to type a text with 10 characters at least',
-         'type': 'string',
+         'type': <any> 'string',
          'default': '',
          'minLength': 10
       },
       'maxLengthText': {
          'title': 'Text with a max length',
          'description': 'You have to type a text with less than 20 characters',
-         'type': 'string',
+         'type': <any> 'string',
          'default': '',
          'maxLength': 20
       },
       'minAndMaxLengthText': {
          'title': 'Text with a min and max length',
          'description': 'You have to type a text with less than 20 and 10 characters at least',
-         'type': 'string',
+         'type': <any> 'string',
          'default': '',
          'minLength': 10,
          'maxLength': 20
@@ -118,36 +119,36 @@ export const JSON_SCHEMA: JSONSchema4 = {
       'url': {
          'title': 'URL',
          'description': 'You have to type a valid url',
-         'type': 'string',
+         'type': <any> 'string',
          'pattern': '(https?:\\/\\/(?:www\\.|(?!www))[^\\s\\.]+\\.[^\\s]{2,}|www\\.[^\\s]+\\.[^\\s]{2,})'
       },
       'boolean': {
          'title': 'Enable or disable this property',
          'description': 'You can enable or disable this property',
-         'type': 'boolean',
+         'type': <any> 'boolean',
          'default': true
       },
       'noDescription': {
          'title': 'Property without description',
-         'type': 'string'
+         'type': <any> 'string'
       },
       'log_level': {
          'title': 'Log Level',
          'description': 'Set the log level: TRACE,DEBUG,INFO,WARN,ERROR and FATAL',
-         'type': 'string',
+         'type': <any> 'string',
          'default': 'INFO',
          'enum': ['TRACE', 'DEBUG', 'INFO', 'WARN', 'ERROR', 'FATAL']
       },
       'executor': {
          'title': 'Executor',
-         'type': 'object',
+         'type': <any> 'object',
          'description': 'Spark executor properties',
          'optional': true,
          'properties': {
             'cores': {
                'title': 'Cores',
                'description': 'Number of per spark`s executor',
-               'type': 'integer',
+               'type': <any> 'integer',
                'default': 4,
                'minimumExclusive': 0,
                'maximum': 8
@@ -155,7 +156,7 @@ export const JSON_SCHEMA: JSONSchema4 = {
             'memory': {
                'title': 'Memory',
                'description': 'Spark executor memory',
-               'type': 'integer',
+               'type': <any> 'integer',
                'default': 512,
                'minimum': 512,
                'maximum': 4096
@@ -163,21 +164,21 @@ export const JSON_SCHEMA: JSONSchema4 = {
             'home': {
                'title': 'Executor home',
                'description': 'Set the directory in which Spark is installed on the executors in Mesos.',
-               'type': 'string',
+               'type': <any> 'string',
                'default': '/opt/spark/dist'
             },
             'subexecutor': {
                'title': 'Sub executor',
-               'type': 'object',
+               'type': <any> 'object',
                'description': 'Spark executor properties',
                'properties': {
                   'subcores': {
                      'title': 'Sub cores',
-                     'type': 'integer'
+                     'type': <any> 'integer'
                   },
                   'submemory': {
                      'title': 'Sub memory',
-                     'type': 'integer'
+                     'type': <any> 'integer'
                   }
                }
             }